draws ALL the flowers on top of eachother
